:white_check_mark: textual descriptions and selectors extraction
