+++ test/constants/selectors.js
@@ -1,0 +1,132 @@
+const ss = {
+  app: '#app',
+  transactoinsTable: '.transaction-results',
+  transactionRow: '.transactions-row',
+  seeAllTxsBtn: '.seeAllLink',
+  txDetailsBackButton: '.transaction-details-back-button',
+  recipientInput: '.recipient input',
+  accountInfoTab: '.account-info',
+  delegateStatisticsTab: '.delegate-statistics ',
+  votedAddress: '.votes .voter-address',
+  voterAdress: '.voters .voter-address',
+  delegateName: '.delegate-name',
+  accountAddress: '.copy-title',
+  searchInput: '#autosuggest-input',
+  showMoreVotesBtn: '.show-votes',
+  leftBlockAccountExplorer: '.explorer-account-left-block',
+  amountInput: '.amount input',
+  bookmarkInput: '#bookmark-input',
+  bookmarkList: '.bookmarkList',
+  nextTransferBtn: '.send-next-button',
+  sendBtn: '.send-button',
+  transactionAddress: '.transaction-address span',
+  chooseDelegateName: '.choose-name',
+  delegateNameInput: '.delegate-name input',
+  submitDelagateNameBtn: '.submit-delegate-name',
+  successText: '.success-description',
+  goToDashboardAfterDelegateReg: '.registration-success',
+  confirmDelegateRegBtn: '.confirm-delegate-registration',
+  spinner: '.spinner',
+  transactionReference: '.transaction-reference',
+  transactionAmount: '.transactionAmount span',
+  transactionAmountPlaceholder: '.transactionAmount',
+  delegateDuplicateNameError: '.error-name-duplicate',
+  headerBalance: '.balance span',
+  nextBtn: '.next',
+  txHeader: '.tx-header',
+  txSenderAddress: '.sender-address',
+  txRecipientAddress: '.receiver-address',
+  txDatePlaceholder: '.tx-date',
+  txDate: '.tx-date .date',
+  txTime: '.tx-time .time',
+  txAddedVotes: '.tx-added-votes .voter-address',
+  txRemovedVotes: '.tx-removed-votes .voter-address',
+  txAmount: '.tx-amount .transactionAmount',
+  txFee: '.tx-fee span',
+  txConfirmations: '.tx-confirmation',
+  txId: '.tx-id .copy-title',
+  txReference: '.tx-reference',
+  delegateResults: '.delegates-result',
+  transactionResults: '.transactions-result',
+  recentSearches: '.addresses-result',
+  idResults: '.addresses-result',
+  closeSearchBtn: '.autosuggest-btn-close',
+  accountBalance: '.balance span',
+  transactionId: '.transaction-id .copy-title',
+  logoutBtn: '.logout',
+  dialogButtons: '.ok-button',
+  emptyResultsMessage: '.empty-message',
+  revealCheckbox: '.reveal-checkbox',
+  passphraseTextarea: 'textarea.passphrase',
+  itsSafeBtn: '.yes-its-safe-button',
+  passphraseWordHolder: '.passphrase-holder label',
+  getToDashboardBtn: '.get-to-your-dashboard-button',
+  confirmCheckbox: '.confirm-checkbox',
+  sidebarMenuSettingsBtn: '#settings',
+  registerSecondPassphraseBtn: '.register-second-passphrase',
+  secondPassphraseIsRegisteredLabel: '.second-passphrase-registered',
+  currencyUSDBtn: '.currency-USD',
+  currencyEURBtn: '.currency-EUR',
+  autoLogoutTrigger: '.autoLog',
+  switchNetworksTrigger: '.showNetwork',
+  delegateFeaturesTrigger: '.advancedMode',
+  messageInput: 'textarea',
+  resulteBtn: '.result',
+  sidebarMenuWalletBtn: '#transactions',
+  referenceInput: '.reference input',
+  convertorElement: '.convertor',
+  convertedPrice: '.converted-price',
+  secondPassphraseInput: '.second-passphrase input',
+  secondPassphraseNextBtn: '.second-passphrase-next',
+  resultMessage: '.result-box-message',
+  okayBtn: '.okay-button',
+  accountInitializationMsg: '.account-initialization',
+  accountInitializationBtn: '.account-init-button',
+  sidebarMenuDelegatesBtn: '#delegates',
+  becomeDelegateLink: '.register-delegate',
+  votesConfirmSidebar: '.confirm-votes',
+  confirmBtn: '.confirm',
+  delegateRow: '.delegate-row',
+  delegateList: '.delegate-list',
+  delegateRank: '.delegate-rank',
+  delegateId: '.delegate-id',
+  delegateProductivity: '.delegate-productivity',
+  searchDelegateInput: 'input.search',
+  voteCheckbox: '.vote-checkbox',
+  voteResultHeader: '.result-box-header',
+  clearSearchBtn: '.clean-icon',
+  votesPreselection: '.upvotes-message',
+  unvotesPreselection: '.unvotes-message',
+  alreadyVotedPreselection: '.alreadyVoted-message',
+  addFollowerAccountButton: '.add-account-button',
+  addressInput: '.address input',
+  followedAccountItem: '.followed-account',
+  titleInput: '.account-title input',
+  followedAccountBalance: '.followed-account-balance',
+  sidebarMenuHelpBtn: '#help',
+  startOnBoardingLink: '.help-onboarding',
+  onBoardingTooltipPrimaryBtn: '.joyride-tooltip__button--primary',
+  onBoardingSkipBtn: '.joyride-tooltip__button--skip',
+  onBoardingHeader: '.joyride-tooltip__header',
+  createLiskIdBtn: '.new-account-button',
+  passphraseInput: '.passphrase input',
+  loginBtn: '.login-button',
+  networkDropdown: '.network',
+  headerAddress: '.copy-title',
+  nodeAddress: '.peer',
+  networkStatus: '.network-status',
+  errorPopup: '.toast',
+  newsBlock: '.news-item',
+  editNewsFeed: '.settingsButton',
+  settingsNewsFeedBlock: '.settingsNewsFeed',
+  getPassphraseButton: '.get-passphrase-button',
+  iUnderstandCheckbox: '.i-understand-checkbox',
+  editFollowingAccounts: '.edit-accounts',
+  removeFollowingAccount: '.remove-account',
+  followedAccount: '.followed-account',
+  followedAccountTitle: '.account-title input',
+  takeTutorial: '.help-onboarding',
+  tutorialTooltip: '.joyride-tooltip__header',
+};
+
+export default ss;
