+++ test/cypress/e2e/bookmarks.spec.js
@@ -1,19 +1,13 @@
 import accounts from '../../constants/accounts';
 import networks from '../../constants/networks';
+import ss from '../../constants/selectors';
 import urls from '../../constants/urls';
 
-const ss = {
-  recipientInput: '.recipient input',
-  amountInput: '.amount input',
-  bookmarkInput: '#bookmark-input',
-  bookmarkList: '.bookmarkList',
-  nextButton: '.send-next-button',
-  sendButton: '.send-button',
-  transactionRow: '.transactions-row',
-  transactionAddress: '.transaction-address span',
-};
-
 describe('Bookmarks', () => {
+  /**
+   * Bookmarks suggestions are not present if there is no followers
+   * @expect bookmarks components are not present
+   */
   it('Bookmarks are not present if there is no followers', () => {
     cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
     cy.visit(urls.wallet);
@@ -22,6 +16,13 @@ describe('Bookmarks', () => {
     cy.get(ss.bookmarkList).should('not.exist');
   });
 
+  /**
+   * Choose follower from bookmarks and send tx
+   * @expect bookmark contain name
+   * @expect bookmark contain address
+   * @expect clicking bookmark fills recipient
+   * @expect tx appears in activity with right address
+   */
   it('Choose follower from bookmarks and send tx', () => {
     window.localStorage.setItem('followedAccounts', `[
       {"title":"Alice","address":"${accounts.delegate.address}","balance":101}
@@ -35,11 +36,17 @@ describe('Bookmarks', () => {
     cy.get(ss.bookmarkList).eq(0).click();
     cy.get(ss.recipientInput).should('have.value', accounts.delegate.address);
     cy.get(ss.amountInput).click().type(1);
-    cy.get(ss.nextButton).click();
-    cy.get(ss.sendButton).click();
+    cy.get(ss.nextTransferBtn).click();
+    cy.get(ss.sendBtn).click();
     cy.get(ss.transactionRow).eq(0).find(ss.transactionAddress).should('have.text', accounts.delegate.address);
   });
 
+  /**
+   * Search through bookmarks by typing
+   * @expect account found by name
+   * @expect non-existent search show empty list
+   * @expect account found by delegate
+   */
   it('Search through bookmarks by typing', () => {
     window.localStorage.setItem('followedAccounts', `[
       {"title":"Alice","address":"${accounts.delegate.address}","balance":101},
