+++ test/cypress/e2e/help.spec.js
@@ -1,32 +1,35 @@
 import accounts from '../../constants/accounts';
 import networks from '../../constants/networks';
 import urls from '../../constants/urls';
+import ss from '../../constants/selectors';
 
-const ss = {
-  sidebarMenuHelpBtn: '#help',
-  startOnBoardingLink: '.help-onboarding',
-  onBoardingTooltipPrimaryBtn: '.joyride-tooltip__button--primary',
-  onBoardingSkipBtn: '.joyride-tooltip__button--skip',
-  onBoardingHeader: '.joyride-tooltip__header',
-};
-
-const checkHelpPageLoaded = () => cy.get('.help-onboarding');
+const checkHelpPageLoaded = () => cy.get(ss.takeTutorial);
 
 const getSettingsObjFromLS = () => JSON.parse(localStorage.getItem('settings'));
 
 describe('Help', () => {
+  /**
+   * Help page can be opened by direct link
+   * @expect url is correct
+   * @expect some specific to page element is present on it
+   */
   it(`opens by url ${urls.help}`, () => {
     cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
     cy.visit(urls.help);
-    cy.url().should('contain', 'help');
+    cy.url().should('contain', urls.help);
     checkHelpPageLoaded();
   });
 
+  /**
+   * Help page can be opened clicking sidebar button
+   * @expect url is correct
+   * @expect some specific to page element is present on it
+   */
   it('opens by sidebar button', () => {
     cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
     cy.visit('/');
     cy.get(ss.sidebarMenuHelpBtn).should('have.css', 'opacity', '1').click();
-    cy.url().should('contain', 'help');
+    cy.url().should('contain', urls.help);
     checkHelpPageLoaded();
   });
 
@@ -35,19 +38,30 @@ describe('Help', () => {
       cy.clearLocalStorage(); // To remove onBoarding: false
     });
 
+    /**
+     * Tutorial does not start on app start
+     */
     it('does not start when not logged in, onBoarding = true', () => {
       cy.addLocalStorage('settings', 'onBoarding', true);
       cy.visit('/');
-      cy.get('.joyride-tooltip__header').should('not.exist');
+      cy.get(ss.tutorialTooltip).should('not.exist');
     });
 
+    /**
+     * Tutorial does not start after sign in if onBoarding is set to false
+     */
     it('does not start when logged in, onBoarding = false', () => {
       cy.addLocalStorage('settings', 'onBoarding', false);
       cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
       cy.visit('/');
-      cy.get('.joyride-tooltip__header').should('not.exist');
+      cy.get(ss.tutorialTooltip).should('not.exist');
     });
 
+    /**
+     * Tutorial start after sign in if onBoarding is not set
+     * Go through tutorial
+     * @expect onBoarding is set to false
+     */
     it('pops up on clean login, go through onboarding', () => {
       cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
       cy.visit('/');
@@ -72,15 +86,24 @@ describe('Help', () => {
         .and(() => expect(getSettingsObjFromLS().onBoarding).to.equal(false));
     });
 
+    /**
+     * Link for tutorial is not there if not logged in
+     * @expect no start tutorial link
+     */
     it('link for onboarding is not there if not logged in', () => {
       cy.visit(urls.help);
       cy.get(ss.startOnBoardingLink).should('not.exist');
     });
 
+    /**
+     * Link for tutorial is there if not logged in
+     * Go through tutorial
+     * @expect onBoarding is set to false
+     */
     it('link is there if logged in, go through onboarding', () => {
       cy.addLocalStorage('settings', 'onBoarding', false);
       cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
-      cy.visit('/help');
+      cy.visit(urls.help);
       cy.get(ss.startOnBoardingLink).click();
       cy.get(ss.onBoardingHeader).should('have.text', 'Welcome to Lisk Hub');
       cy.get(ss.onBoardingTooltipPrimaryBtn).click();
@@ -102,6 +125,10 @@ describe('Help', () => {
         .and(() => expect(getSettingsObjFromLS().onBoarding).to.equal(false));
     });
 
+    /**
+     * Skip tutorial in the middle
+     * @expect onBoarding is set to false
+     */
     it('skip onboarding in the process', () => {
       cy.addLocalStorage('settings', 'onBoarding', true);
       cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
