+++ test/cypress/e2e/login.spec.js
@@ -2,19 +2,9 @@ import numeral from 'numeral';
 import { fromRawLsk } from '../../../src/utils/lsk';
 import accounts from '../../constants/accounts';
 import networks from '../../constants/networks';
+import ss from '../../constants/selectors';
+import urls from '../../constants/urls';
 import chooseNetwork from '../utils/chooseNetwork';
-
-const ss = {
-  newAccountBtn: '.new-account-button',
-  passphraseInput: '.passphrase input',
-  loginBtn: '.login-button',
-  networkDropdown: '.network',
-  headerAddress: '.copy-title',
-  headerBalance: '.balance span',
-  nodeAddress: '.peer',
-  networkStatus: '.network-status',
-  errorPopup: '.toast',
-};
 
 const loginUI = (passphrase) => {
   cy.get(ss.passphraseInput).click();
@@ -28,13 +18,23 @@ const loginUI = (passphrase) => {
 const castNumberToBalanceString = number => numeral(fromRawLsk(number)).format('0,0.[0000000000000]');
 
 describe('Login Page', () => {
+  /**
+   * 'Create lisk id' link leads to registration page
+   * @expect some specific to page element is present on it
+   */
   it('Create lisk id -> Register account page', () => {
     cy.visit('/');
-    cy.get(ss.newAccountBtn).click();
-    cy.url().should('include', '/register');
-    cy.get('.multistep-back');
+    cy.get(ss.createLiskIdBtn).click();
+    cy.url().should('include', urls.register);
+    cy.get(ss.app).contains('Create your Lisk ID');
   });
 
+  /**
+   * First time app is opened sign in happens to mainnet
+   * @expect address matches
+   * @expect 0 balance - otherwise you've just found free money
+   * @expect network status is not shown
+   */
   it('Log in to Mainnet by default ("Switch Network" is not set)', () => {
     cy.visit('/');
     loginUI(accounts.genesis.passphrase);
@@ -43,6 +43,12 @@ describe('Login Page', () => {
     cy.get(ss.networkStatus).should('not.exist');
   });
 
+  /**
+   * If the switch network is set to false sign in happens to mainnet
+   * @expect address matches
+   * @expect 0 balance - otherwise you've just found free money
+   * @expect network status is not shown
+   */
   it('Log in to Mainnet by default ("Switch Network" is false)', () => {
     cy.addLocalStorage('settings', 'showNetwork', false);
     cy.visit('/');
@@ -52,6 +58,12 @@ describe('Login Page', () => {
     cy.get(ss.networkStatus).should('not.exist');
   });
 
+  /**
+   * If the switch network is set to true and mainnet is chosen sign in happens to mainnet
+   * @expect address matches
+   * @expect 0 balance - otherwise you've just found free money
+   * @expect network status shows mainnet
+   */
   it('Log in to Mainnet (Selected network)', () => {
     cy.addLocalStorage('settings', 'showNetwork', true);
     cy.visit('/');
@@ -62,6 +74,12 @@ describe('Login Page', () => {
     cy.get(ss.headerBalance).should('have.text', castNumberToBalanceString(0));
   });
 
+  /**
+   * If the switch network is set to true and testnet is chosen sign in happens to testnet
+   * @expect address matches
+   * @expect balance matches one specified in constants therefore please don't touch it :)
+   * @expect network status shows testnet
+   */
   it('Log in to Testnet', () => {
     cy.addLocalStorage('settings', 'showNetwork', true);
     cy.visit('/');
@@ -72,6 +90,13 @@ describe('Login Page', () => {
     cy.get(ss.headerBalance).should('have.text', castNumberToBalanceString(accounts['testnet guy'].balance));
   });
 
+  /**
+   * If the switch network is set to true, devnet is chosen and custom node is specified,
+   * sign in happens to testnet
+   * @expect address matches
+   * @expect balance's three leading numbers match specified in constants
+   * @expect network status shows devnet
+   */
   it('Log in to Devnet', () => {
     cy.addLocalStorage('settings', 'showNetwork', true);
     cy.visit('/');
@@ -83,11 +108,19 @@ describe('Login Page', () => {
     cy.get(ss.nodeAddress).contains(networks.devnet.node);
   });
 
+  /**
+   * ?showNetwork parameter makes network switcher available
+   * @expect network switcher is visible
+   */
   it('Network switcher available by url ?showNetwork=true', () => {
     cy.visit('?showNetwork=true');
     cy.get(ss.networkDropdown).should('be.visible');
   });
 
+  /**
+   * Try to sign in to invalid custom node address
+   * @expect error popup is shown
+   */
   it('Log in to invalid address', () => {
     cy.addLocalStorage('settings', 'showNetwork', true);
     cy.visit('/');
