+++ test/cypress/e2e/registration.spec.js
@@ -1,23 +1,9 @@
 import networks from '../../constants/networks';
 import urls from '../../constants/urls';
+import ss from '../../constants/selectors';
 import chooseNetwork from '../utils/chooseNetwork';
 import moveMouseRandomly from '../utils/moveMouseRandomly';
 import slideCheckbox from '../utils/slideCheckbox';
-
-const ss = {
-  networkStatus: '.network-status',
-  newAccountBtn: '.new-account-button',
-  nodeAddress: '.peer',
-  errorPopup: '.toast',
-  getPassphraseButton: '.get-passphrase-button',
-  iUnderstandCheckbox: '.i-understand-checkbox',
-  revealCheckbox: '.reveal-checkbox',
-  passphraseTextarea: 'textarea.passphrase',
-  itsSafeBtn: '.yes-its-safe-button',
-  passphraseWordHolder: '.passphrase-holder label',
-  getToDashboardBtn: '.get-to-your-dashboard-button',
-  backButton: '.multistep-back',
-};
 
 const registerUI = function () {
   moveMouseRandomly();
@@ -33,18 +19,31 @@ const registerUI = function () {
 };
 
 describe('Registration', () => {
+  /**
+   * Registration page can be opened by direct link
+   * @expect url is correct
+   * @expect some specific to page element is present on it
+   */
   it(`Opens by url ${urls.register}`, () => {
     cy.visit(urls.register);
     cy.url().should('contain', 'register');
-    cy.get(ss.backButton);
+    cy.get(ss.app).contains('Create your Lisk ID');
   });
 
+  /**
+   * Create Lisk ID for Mainnet by default ("Switch Network" is not set)
+   * @expect network status is not shown
+   */
   it('Create Lisk ID for Mainnet by default ("Switch Network" is not set)', function () {
     cy.visit(urls.register);
     registerUI.call(this);
     cy.get(ss.networkStatus).should('not.exist');
   });
 
+  /**
+   * Create Lisk ID for Mainnet by default ("Switch Network" is false)
+   * @expect network status is not shown
+   */
   it('Create Lisk ID for Mainnet ("Switch Network" is false)', function () {
     cy.addLocalStorage('settings', 'showNetwork', false);
     cy.visit(urls.register);
@@ -52,40 +51,56 @@ describe('Registration', () => {
     cy.get(ss.networkStatus).should('not.exist');
   });
 
-  it('Create Lisk ID for Mainnet (Selected network)', function () {
+  /**
+   * Create Lisk ID for Mainnet
+   * @expect network status shows mainnet
+   */
+  it('Create Lisk ID for Mainnet', function () {
     cy.addLocalStorage('settings', 'showNetwork', true);
     cy.visit('/');
     chooseNetwork('main');
-    cy.get(ss.newAccountBtn).click();
+    cy.get(ss.createLiskIdBtn).click();
     registerUI.call(this);
     cy.get(ss.networkStatus).contains('Connected to mainnet');
   });
 
+  /**
+   * Create Lisk ID for Testnet
+   * @expect network status shows testnet
+   */
   it('Create Lisk ID for Testnet', function () {
     cy.addLocalStorage('settings', 'showNetwork', true);
     cy.visit('/');
     chooseNetwork('test');
-    cy.get(ss.newAccountBtn).click();
+    cy.get(ss.createLiskIdBtn).click();
     registerUI.call(this);
     cy.get(ss.networkStatus).contains('Connected to testnet');
   });
 
+  /**
+   * Create Lisk ID for Devnet
+   * @expect network status shows devnet
+   */
   it('Create Lisk ID for Devnet', function () {
     cy.addLocalStorage('settings', 'showNetwork', true);
     cy.visit('/');
     chooseNetwork('dev');
-    cy.get(ss.newAccountBtn).click();
+    cy.get(ss.createLiskIdBtn).click();
     registerUI.call(this);
     cy.get(ss.networkStatus).contains('Connected to devnet');
     cy.get(ss.nodeAddress).contains(networks.devnet.node);
   });
 
+  /**
+   * Try Create Lisk ID with invalid address entered
+   * @expect error shown
+   */
   it('Create Lisk ID for invalid address', () => {
     cy.addLocalStorage('settings', 'showNetwork', true);
     cy.visit('/');
     chooseNetwork('invalid');
-    cy.get(ss.newAccountBtn).click();
+    cy.get(ss.createLiskIdBtn).click();
     cy.get(ss.errorPopup).contains('Unable to connect to the node');
-    cy.get(ss.newAccountBtn);
+    cy.get(ss.createLiskIdBtn);
   });
 });
