+++ test/cypress/e2e/search.spec.js
@@ -1,28 +1,11 @@
 import accounts from '../../constants/accounts';
 import networks from '../../constants/networks';
-
-const ss = {
-  searchInput: '#autosuggest-input',
-  delegateResults: '.delegates-result',
-  transactionResults: '.transactions-result',
-  recentSearches: '.addresses-result',
-  idResults: '.addresses-result',
-  clearSearchBtn: '.autosuggest-btn-close',
-  explorerAccountLeftBlock: '.explorer-account-left-block',
-  accountAddress: '.copy-title',
-  accountBalance: '.balance span',
-  delegateName: '.delegate-name',
-  transactionId: '.transaction-id .copy-title',
-  logoutBtn: '.logout',
-  dialogButtons: '.ok-button',
-  emptyResultsMessage: '.empty-message',
-};
-
-const testnetTransaction = '755251579479131174';
+import ss from '../../constants/selectors';
 
 const getSearchesObjFromLS = () => JSON.parse(localStorage.getItem('searches'));
 
 describe('Search', () => {
+  const testnetTransaction = '755251579479131174';
   [
     {
       name: 'Search for Lisk ID using keyboard Enter',
@@ -38,11 +21,17 @@ describe('Search', () => {
       },
     },
   ].forEach((testSet) => {
+    /**
+     * Search for Lisk ID
+     * @expect account page with corresponding ID is a result
+     * @expect localStorage have the searches object with correct address
+     * @expect localStorage have the searches object with correct searchTerm
+     */
     it(testSet.name, () => {
       cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
       cy.visit('/');
       testSet.searchAction();
-      cy.get(ss.explorerAccountLeftBlock).find(ss.accountAddress).should('have.text', accounts.delegate.address)
+      cy.get(ss.leftBlockAccountExplorer).find(ss.accountAddress).should('have.text', accounts.delegate.address)
         .and(() => {
           expect(getSearchesObjFromLS()[0].id).to.equal(accounts.delegate.address);
           expect(getSearchesObjFromLS()[0].searchTerm).to.equal(accounts.delegate.address);
@@ -65,6 +54,12 @@ describe('Search', () => {
       },
     },
   ].forEach((testSet) => {
+    /**
+     * Search for transaction
+     * @expect transaction details page a result
+     * @expect localStorage have the searches object with correct id
+     * @expect localStorage have the searches object with correct searchTerm
+     */
     it(testSet.name, () => {
       cy.autologin(accounts.genesis.passphrase, networks.testnet.node);
       cy.visit('/');
@@ -93,11 +88,17 @@ describe('Search', () => {
       },
     },
   ].forEach((testSet) => {
+    /**
+     * Search for delegate
+     * @expect account page with corresponding ID is a result
+     * @expect localStorage have the searches object with correct address
+     * @expect localStorage have the searches object with correct searchTerm
+     */
     it(testSet.name, () => {
       cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
       cy.visit('/');
       testSet.searchAction();
-      cy.get(ss.explorerAccountLeftBlock).find(ss.delegateName).should('have.text', accounts.delegate.username)
+      cy.get(ss.leftBlockAccountExplorer).find(ss.delegateName).should('have.text', accounts.delegate.username)
         .and(() => {
           expect(getSearchesObjFromLS()[0].id).to.equal(accounts.delegate.address);
           expect(getSearchesObjFromLS()[0].searchTerm).to.equal(accounts.delegate.username);
@@ -105,19 +106,31 @@ describe('Search', () => {
     });
   });
 
+  /**
+   * Search without signing in
+   * @expect happens in mainnet
+   */
   it('Search without signing in - happens in mainnet', () => {
     cy.visit('/');
     cy.get(ss.searchInput).click().type(`${accounts['mainnet delegate'].address}{enter}`);
-    cy.get(ss.explorerAccountLeftBlock).find(ss.delegateName).should('have.text', accounts['mainnet delegate'].username);
+    cy.get(ss.leftBlockAccountExplorer).find(ss.delegateName).should('have.text', accounts['mainnet delegate'].username);
   });
 
+  /**
+   * Search signed in mainnet
+   * @expect happens in mainnet
+   */
   it('Search signed in mainnet - happens in mainnet', () => {
     cy.autologin(accounts.genesis.passphrase, networks.mainnet.node);
     cy.visit('/');
     cy.get(ss.searchInput).click().type(`${accounts['mainnet delegate'].address}{enter}`);
-    cy.get(ss.explorerAccountLeftBlock).find(ss.delegateName).should('have.text', accounts['mainnet delegate'].username);
+    cy.get(ss.leftBlockAccountExplorer).find(ss.delegateName).should('have.text', accounts['mainnet delegate'].username);
   });
 
+  /**
+   * Search signed in testnet
+   * @expect happens in testnet
+   */
   it('Search signed in testnet - happens in testnet', () => {
     cy.autologin(accounts.genesis.passphrase, networks.testnet.node);
     cy.visit('/');
@@ -126,13 +139,21 @@ describe('Search', () => {
     cy.get(ss.transactionId).should('have.text', testnetTransaction);
   });
 
+  /**
+   * Search signed in devnet
+   * @expect happens in devnet
+   */
   it('Search signed in devnet - happens in devnet', () => {
     cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
     cy.visit('/');
     cy.get(ss.searchInput).click().type(`${accounts.delegate.address}{enter}`);
-    cy.get(ss.explorerAccountLeftBlock).find(ss.delegateName).should('have.text', accounts.delegate.username);
+    cy.get(ss.leftBlockAccountExplorer).find(ss.delegateName).should('have.text', accounts.delegate.username);
   });
 
+  /**
+   * Search after logout
+   * @expect happens in last used network
+   */
   it('Search after logout - happens in last used network', () => {
     cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
     cy.visit('/');
@@ -141,18 +162,27 @@ describe('Search', () => {
     cy.get(ss.dialogButtons).eq(1).click();
     cy.get(ss.searchInput).click().type(`${accounts.delegate.username}`);
     cy.get(ss.delegateResults).eq(0).click();
-    cy.get(ss.explorerAccountLeftBlock).find(ss.delegateName).should('have.text', accounts.delegate.username);
+    cy.get(ss.leftBlockAccountExplorer).find(ss.delegateName).should('have.text', accounts.delegate.username);
   });
 
+  /**
+   * Recent searches are shown as search proposals and clickable
+   * @expect recent searches are shown
+   * @expect click on proposal leads to corresponding page
+   */
   it('Recent searches are shown as search proposals and clickable', () => {
     cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
     window.localStorage.setItem('searches', `[{"searchTerm":"${accounts.genesis.address}","id":"${accounts.genesis.address}"}]`);
     cy.visit('/');
     cy.get(ss.searchInput).click();
     cy.get(ss.recentSearches).eq(0).click();
-    cy.get(ss.explorerAccountLeftBlock).find(ss.accountAddress).should('have.text', accounts.genesis.address);
+    cy.get(ss.leftBlockAccountExplorer).find(ss.accountAddress).should('have.text', accounts.genesis.address);
   });
 
+  /**
+   * Search for nonexistent item
+   * @expect no results message
+   */
   it('Search for nonexistent item - shows no results message', () => {
     cy.visit('/');
     cy.get(ss.searchInput).click().type('43th3j4bt324{enter}');
