+++ test/cypress/e2e/transfer.spec.js
@@ -1,32 +1,9 @@
 import accounts from '../../constants/accounts';
 import networks from '../../constants/networks';
 import urls from '../../constants/urls';
+import ss from '../../constants/selectors';
 import enterSecondPassphrase from '../utils/enterSecondPassphrase';
 import compareBalances from '../utils/compareBalances';
-
-const ss = {
-  sidebarMenuWalletBtn: '#transactions',
-  recipientInput: '.recipient input',
-  amountInput: '.amount input',
-  referenceInput: '.reference input',
-  convertorElement: '.convertor',
-  convertedPrice: '.converted-price',
-  nextButton: '.send-next-button',
-  sendButton: '.send-button',
-  secondPassphraseInput: '.second-passphrase input',
-  secondPassphraseNextBtn: '.second-passphrase-next',
-  resultMessage: '.result-box-message',
-  okayBtn: '.okay-button',
-  transactoinsTable: '.transaction-results',
-  transactionRow: '.transactions-row',
-  accountInitializationMsg: '.account-initialization',
-  accountInitializationBtn: '.account-init-button',
-  spinner: '.spinner',
-  transactionAddress: '.transaction-address span',
-  transactionReference: '.transaction-reference',
-  transactionAmount: '.transactionAmount span',
-  headerBalance: '.balance span',
-};
 
 const msg = {
   transferTxSuccess: 'Transaction is being processed and will be confirmed. It may take up to 15 minutes to be secured in the blockchain.',
@@ -54,21 +31,39 @@ describe('Transfer', () => {
     randomReference = getRandomReference();
   });
 
+  /**
+   * Wallet page can be opened by direct link
+   * @expect url is correct
+   * @expect some specific to page element is present on it
+   */
   it(`Wallet page opens by url ${urls.wallet}`, () => {
     cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
     cy.visit(urls.wallet);
-    cy.url().should('contain', 'wallet');
+    cy.url().should('contain', urls.wallet);
     checkWalletPageLoaded();
   });
 
+  /**
+   * Sidebar link leads to Wallet page
+   * @expect url is correct
+   * @expect some specific to page element is present on it
+   */
   it('Wallet page opens by sidebar button', () => {
     cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
     cy.visit(urls.dashboard);
     cy.get(ss.sidebarMenuWalletBtn).should('have.css', 'opacity', '1').click();
-    cy.url().should('contain', 'wallet');
+    cy.url().should('contain', urls.wallet);
     checkWalletPageLoaded();
   });
 
+  /**
+   * Make a transfer with empty reference
+   * @expect successfully go through transfer process
+   * @expect transaction appears in the activity list as pending
+   * @expect transaction appears in the activity list with correct data
+   * @expect transaction appears in the activity list as confirmed
+   * @expect header balance value is decreased
+   */
   it('Transfer tx with empty ref appears in activity pending -> approved,' +
     'Header balance is affected', function () {
     cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
@@ -76,8 +71,8 @@ describe('Transfer', () => {
     cy.get(ss.headerBalance).invoke('text').as('balanceBefore');
     cy.get(ss.recipientInput).type(randomAddress);
     cy.get(ss.amountInput).click().type(randomAmount);
-    cy.get(ss.nextButton).click();
-    cy.get(ss.sendButton).click();
+    cy.get(ss.nextTransferBtn).click();
+    cy.get(ss.sendBtn).click();
     cy.get(ss.resultMessage).should('have.text', msg.transferTxSuccess);
     cy.get(ss.transactionRow).eq(0).as('tx');
     cy.get('@tx').find('.spinner');
@@ -91,14 +86,21 @@ describe('Transfer', () => {
     });
   });
 
+  /**
+   * Make a transfer with reference
+   * @expect successfully go through transfer process
+   * @expect transaction appears in the activity list as pending
+   * @expect transaction appears in the activity list with correct data
+   * @expect transaction appears in the activity list as confirmed
+   */
   it('Transfer tx with ref appears in dashboard activity pending -> approved', () => {
     cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
     cy.visit(urls.wallet);
     cy.get(ss.recipientInput).type(randomAddress);
     cy.get(ss.referenceInput).click().type(randomReference);
     cy.get(ss.amountInput).click().type(randomAmount);
-    cy.get(ss.nextButton).click();
-    cy.get(ss.sendButton).click();
+    cy.get(ss.nextTransferBtn).click();
+    cy.get(ss.sendBtn).click();
     cy.get(ss.resultMessage).should('have.text', msg.transferTxSuccess);
     cy.visit(urls.dashboard);
     cy.get(ss.transactionRow).eq(0).as('tx');
@@ -110,49 +112,74 @@ describe('Transfer', () => {
     cy.get('@tx').find(ss.spinner).should('not.exist');
   });
 
+  /**
+   * Make a transfer with second passphrase
+   * @expect successfully go through transfer process
+   * @expect transaction appears in the activity list as pending
+   * @expect transaction appears in the activity list with correct data
+   * @expect transaction appears in the activity list as confirmed
+   */
   it('Transfer tx with second passphrase', () => {
     cy.autologin(accounts['second passphrase account'].passphrase, networks.devnet.node);
     cy.visit(urls.wallet);
     cy.get(ss.recipientInput).type(randomAddress);
     cy.get(ss.referenceInput).click().type(randomReference);
     cy.get(ss.amountInput).click().type(randomAmount);
-    cy.get(ss.nextButton).click();
+    cy.get(ss.nextTransferBtn).click();
     enterSecondPassphrase(accounts['second passphrase account'].secondPassphrase);
-    cy.get(ss.sendButton).click();
+    cy.get(ss.sendBtn).click();
     cy.get(ss.resultMessage).should('have.text', msg.transferTxSuccess);
     cy.get(ss.transactionRow).eq(0).as('tx');
     cy.get('@tx').find('.spinner');
     cy.get('@tx').find(ss.transactionAddress).should('have.text', randomAddress);
   });
 
+  /**
+   * Shortcut URL prefills recipient, amount and reference
+   * @expect recipient, amount and reference are prefilled
+   */
   it('Launch protocol link prefills recipient, amount and reference', () => {
     cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
-    cy.visit('/wallet?recipient=4995063339468361088L&amount=5&reference=test');
+    cy.visit(`${urls.wallet}/?recipient=4995063339468361088L&amount=5&reference=test`);
     cy.get(ss.recipientInput).should('have.value', '4995063339468361088L');
     cy.get(ss.amountInput).should('have.value', '5');
     cy.get(ss.referenceInput).should('have.value', 'test');
   });
 
+  /**
+   * Fiat converter shows amount in USD if set to USD
+   * @expect amount in USD
+   */
   it('Fiat converter shows amount in USD', () => {
     cy.addLocalStorage('settings', 'currency', 'USD');
     cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
-    cy.visit('/wallet?recipient=4995063339468361088L&amount=5');
+    cy.visit(`${urls.wallet}?recipient=4995063339468361088L&amount=5`);
     cy.get(ss.convertedPrice).contains(/^\d{1,100}(\.\d{1,2})? USD$/);
   });
 
+  /**
+   * Fiat converter shows amount in EUR if set to EUR
+   * @expect amount in EUR
+   */
   it('Fiat converter shows amount in EUR', () => {
     cy.addLocalStorage('settings', 'currency', 'EUR');
     cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
-    cy.visit('/wallet?recipient=4995063339468361088L&amount=5');
+    cy.visit(`${urls.wallet}?recipient=4995063339468361088L&amount=5`);
     cy.get(ss.convertedPrice).contains(/^\d{1,100}(\.\d{1,2})? EUR$/);
   });
 
+  /**
+   * Initialization dialogue and functionality is shown for the fresh account
+   * @expect successfully go through initialization
+   * @expect transfer transaction appear with correct data
+   * @expect initialization dialogue is not shown anymore
+   */
   it('Should be able to init account when needed', () => {
     cy.autologin(accounts['without initialization'].passphrase, networks.devnet.node);
     cy.reload();
     cy.visit(urls.wallet);
     cy.get(ss.accountInitializationMsg).get(ss.accountInitializationBtn).click();
-    cy.get(ss.sendButton).click();
+    cy.get(ss.sendBtn).click();
     cy.get(ss.resultMessage).should('have.text', msg.transferTxSuccess);
     cy.get(ss.accountInitializationMsg).should('not.exist');
     cy.wait(txConfirmationTimeout);
